+++ src/html/options.html
@@ -2,7 +2,8 @@
 <html>
   <head>
     <meta charset="utf-8" />
-    <link rel="stylesheet" type="text/css" href="options.css" />
+    <link rel="stylesheet" type="text/css" href="/src/css/options.css" />
+    <script src="/src/js/private/options.mjs" type="module"></script>
     <title>OdooTerminal Options</title>
   </head>
 
@@ -66,7 +67,7 @@
         <p>
           Run these commands when initialize the terminal (one per line).<br />Be
           careful! changes here can cause unwanted exceptions. Note that the
-          commands will be executed on all pages.
+          commands will be executed on all 'odoo' pages.
         </p>
         <textarea id="init_cmds"></textarea>
       </div>
@@ -75,7 +76,7 @@
         <h4>Terminal Context</h4>
         <p>
           This context will be merged with the "normal" context on terminal
-          operations.
+          operations. In json format.
         </p>
         <textarea id="term_context"></textarea>
       </div>
@@ -85,8 +86,5 @@
         <button type="submit">Save</button>
       </div>
     </form>
-
-    <script src="../globals.js" type="text/javascript"></script>
-    <script src="options.js" type="text/javascript"></script>
   </body>
 </html>
