+++ src/js/shared/content/content_script.mjs
@@ -8,79 +8,71 @@ import {ubrowser} from "../globals.mjs";
 import {SETTING_NAMES} from "../../common/globals.mjs";
 import {sendWindowMessage} from "../../common/utils.mjs";
 
-export class ExtensionContent {
-  constructor() {
-    this.#handleEvents();
+/**
+ * @param {Object} odoo_info - The collected information
+ */
+function updateInstanceContext(odoo_info = {}) {
+  if (typeof odoo_info !== "object") {
+    return;
   }
+  updateContext(odoo_info, {isLoaded: true});
+}
 
-  /**
-   * @param {Object} odoo_info - The collected information
-   */
-  #updateInstanceContext(odoo_info = {}) {
-    if (typeof odoo_info !== "object") {
-      return;
-    }
-    updateContext(odoo_info, {isLoaded: true});
+/**
+ * Listen messages from page context
+ * @param {Object} event
+ */
+function onWindowMessage(event) {
+  // We only accept messages from ourselves
+  if (event.source !== window) {
+    return;
   }
-
-  /**
-   * Listen messages from page context
-   * @param {Object} event
-   */
-  #onWindowMessage(event) {
-    // We only accept messages from ourselves
-    if (event.source !== window) {
-      return;
+  if (event.data.type === "ODOO_TERM_INIT") {
+    var info = event.data.instance_info;
+    updateInstanceContext(info);
+    if (info.isCompatible) {
+      injector(document, getResources());
+    } else if (info.isOdoo) {
+      console.warn("[OdooTerminal] Incompatible server version!");
     }
-    if (event.data.type === "ODOO_TERM_INIT") {
-      var info = event.data.instance_info;
-      this.#updateInstanceContext(info);
-      if (info.isCompatible) {
-        injector(document, getResources());
-      } else if (info.isOdoo) {
-        console.warn("[OdooTerminal] Incompatible server version!");
+  } else if (event.data.type === "ODOO_TERM_START") {
+    // Load Init Commands
+    getStorageSync(SETTING_NAMES).then((items) => {
+      const data = {};
+      for (const config_name in items) {
+        data[config_name] = items[config_name];
       }
-    } else if (event.data.type === "ODOO_TERM_START") {
-      // Load Init Commands
-      getStorageSync(SETTING_NAMES).then((items) => {
-        const data = {};
-        for (const config_name in items) {
-          data[config_name] = items[config_name];
-        }
-        sendWindowMessage(window, "ODOO_TERM_CONFIG", {
-          config: data,
-          info: InstanceContext,
-        });
+      sendWindowMessage(window, "ODOO_TERM_CONFIG", {
+        config: data,
+        info: InstanceContext,
       });
-    }
-  }
-
-  /**
-   * Listen message from extension context
-   * @param {Object} request
-   */
-  #onMessage(request) {
-    if (request.message === "update_odoo_terminal_info") {
-      if (InstanceContext.isLoaded) {
-        this.#updateInstanceContext();
-      } else {
-        injectPageScript(
-          document,
-          "src/js/shared/page/instance_analyzer.mjs",
-          (ev) => {
-            ev.target.parentNode.removeChild(ev.target);
-          }
-        );
-      }
-    } else if (request.message === "toggle_terminal") {
-      if (InstanceContext.isCompatible) {
-        document.getElementById("terminal").dispatchEvent(new Event("toggle"));
-      }
-    }
-  }
-
-  #handleEvents() {
-    window.addEventListener("message", this.#onWindowMessage.bind(this), false);
-    ubrowser.runtime.onMessage.addListener(this.#onMessage.bind(this));
+    });
   }
 }
+
+/**
+ * Listen message from extension context
+ * @param {Object} request
+ */
+function onInternalMessage(request) {
+  if (request.message === "update_odoo_terminal_info") {
+    if (InstanceContext.isLoaded) {
+      updateInstanceContext();
+    } else {
+      injectPageScript(
+        document,
+        "src/js/shared/page/instance_analyzer.mjs",
+        (ev) => {
+          ev.target.parentNode.removeChild(ev.target);
+        }
+      );
+    }
+  } else if (request.message === "toggle_terminal") {
+    if (InstanceContext.isCompatible) {
+      document.getElementById("terminal").dispatchEvent(new Event("toggle"));
+    }
+  }
+}
+
+window.addEventListener("message", onWindowMessage, false);
+ubrowser.runtime.onMessage.addListener(onInternalMessage);
